requirements-dev.txt
-r requirements.txt

requests-mock==1.12.1
pytest
pytest-xdist
